Massive work to make lang/l2 compile with sdcc
While lang/l2 was written for a C compiler with structs, and
presumably some type checking, it misused how #include files
should be used and double-declared some structs and global
variables.

It was a big mess. This commit adds prototypes to all functions
and fixes some apparent errors, and marks file-scoped functions
as "static".