First cut, ASM grammar parser
This parser can recognise a substantial fraction of my assembler
source code. The Z80 opcodes recognised and their allowed operands
are incomplete in many respects.

Expressions are incompletely implemented. I expect the assembler
doesn't allow an expression of the form (3+4) where it could be
confused with (HL). Unary minus is not implemented. Negative
numbers are partially implemented. And so on.